% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KernelPool.R
\name{getDelta}
\alias{getDelta}
\title{Delta Kernel for Binary One-Hot Data}
\usage{
getDelta(geno)
}
\arguments{
\item{geno}{Binary matrix (0s and 1s) representing one-hot encoded data}
}
\value{
Symmetric binary kernel matrix
}
\description{
Computes a kernel matrix for binary one-hot encoded data where the kernel
value is 1 when two samples share the same "hot" bit (active category).
}
\keyword{internal}
