% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KernelPool.R
\name{getExponential}
\alias{getExponential}
\title{Exponential Kernel}
\usage{
getExponential(geno)
}
\arguments{
\item{geno}{Numeric matrix of features}
}
\value{
Exponential kernel matrix
}
\description{
Applies exponential transformation to the product kernel.
}
\keyword{internal}
