% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KernelPool.R
\name{getFstidentity}
\alias{getFstidentity}
\title{First-Order Identity Kernel}
\usage{
getFstidentity(geno)
}
\arguments{
\item{geno}{Numeric matrix of features}
}
\value{
Diagonal kernel matrix
}
\description{
Diagonal kernel using diagonal elements from the product kernel.
}
\keyword{internal}
