% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KernelPool.R
\name{getIdentity}
\alias{getIdentity}
\title{Identity Kernel}
\usage{
getIdentity(geno)
}
\arguments{
\item{geno}{Input matrix (used only for determining dimensions)}
}
\value{
Identity matrix
}
\description{
Returns an identity matrix of appropriate dimensions.
}
\keyword{internal}
