% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KernelPool.R
\name{getProductIdentity}
\alias{getProductIdentity}
\title{Product Identity Kernel}
\usage{
getProductIdentity(geno, weights = 1)
}
\arguments{
\item{geno}{Numeric matrix of features}

\item{weights}{Numeric vector of feature weights}
}
\value{
Regularized product kernel matrix
}
\description{
Product kernel with added identity regularization to ensure positive definiteness.
}
\keyword{internal}
