% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KernelPool.R
\name{getRBF}
\alias{getRBF}
\title{Radial Basis Function (RBF) Kernel}
\usage{
getRBF(geno, sigma = 1)
}
\arguments{
\item{geno}{Numeric matrix of features}

\item{sigma}{Numeric bandwidth parameter}
}
\value{
RBF kernel matrix
}
\description{
Alternative implementation of RBF kernel using direct Euclidean distances.
}
\keyword{internal}
