% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KernelPool.R
\name{getRelu}
\alias{getRelu}
\title{ReLU Kernel}
\usage{
getRelu(geno)
}
\arguments{
\item{geno}{Numeric matrix of features}
}
\value{
ReLU kernel matrix
}
\description{
Applies rectified linear unit (ReLU) transformation to product kernel.
Sets negative values to zero.
}
\keyword{internal}
