% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KernelPool.R
\name{getSigmoid}
\alias{getSigmoid}
\title{Sigmoid Kernel}
\usage{
getSigmoid(geno, sigma = 1)
}
\arguments{
\item{geno}{Numeric matrix of features}

\item{sigma}{Numeric shift parameter}
}
\value{
Sigmoid kernel matrix
}
\description{
Applies hyperbolic tangent transformation to the product kernel.
}
\keyword{internal}
