% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KernelPool.R
\name{getSqidentity}
\alias{getSqidentity}
\title{Squared Identity Kernel}
\usage{
getSqidentity(geno)
}
\arguments{
\item{geno}{Numeric matrix of features}
}
\value{
Diagonal kernel matrix
}
\description{
Diagonal kernel with squared diagonal elements from the product kernel.
}
\keyword{internal}
