% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KernelPool.R
\name{L2.norm}
\alias{L2.norm}
\title{L2 Norm Helper Function}
\usage{
L2.norm(X)
}
\arguments{
\item{X}{Numeric vector}
}
\value{
Scalar L2 norm value
}
\description{
Computes the L2 norm (mean squared magnitude) of a vector.
}
\keyword{internal}
