% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scKNN.R
\name{plot.scKNN.result}
\alias{plot.scKNN.result}
\title{Plot Method for scKNN Results}
\usage{
\method{plot}{scKNN.result}(x, type = c("manhattan", "qq", "variance"), ...)
}
\arguments{
\item{x}{An object of class "scKNN.result"}

\item{type}{Character, type of plot: "manhattan", "qq", or "variance"}

\item{...}{Additional plotting arguments}
}
\description{
Plot Method for scKNN Results
}
